perf: rework the txn API to not save data for read items
this allows us to avoid having to remember the contents of each key
read. the downside is that repeated reads of the same key (A) will each
result in a btree lookup instead of being served from the cache, and (B)
might not be the same with concurrent modifications present (in which
case the txn aborts).

we still guarantee that any writes a txn does will be read back,
even with concurrent modifications