refactor: transaction now take a traits type parameter
doing so allows us to statically configure transactions
to have different size read/write sets. this allows us to
tailor the read/write set sizes for each txn separately.

we do this in a type-safe manner, so that the compiler
allows one to mix and match the same txn class type
w/ different transaction traits on the same txn btree, but
*disallows* mixing different transaction types

this is, of course, done w/ more C++ templates!