tpcc: deal with string arena running out of space
we handle this by deviating from TPC-C on the newest order scan- we
can't actually implement the scan w/o potentially touching an unbounded
number of tuples because we don't have reverse scans implemented.

we work around it by sampling from the oldest orders- this gives us more
or less the same # of operations as if we had reverse scans.  it is
somewhat unfair because we might possibly get better cache locality on
scanning the same set of elements over and over again, but this probably
won't make a difference since order_status is an infrequent txn (and
over the same customer is even *more* infrequent) so core cache lines
pertaining to the order lines will almost be guaranteed to be not in the
cache by the time we execute the txn again.