tpcc: add a fast ID generation for new order
this aims to try and remove one of the major sources of aborts when
running lots of new order transactions. it is not really a valid
implementation, but rather an upper bound on the performance we could
reclaim.

note that our tree could *transactionally* offer a similar feature
(we just have not implemented it yet)